Finalizado

+++ Proyecto/HTML/carrito_modif.php
@@ -3,7 +3,13 @@
 @ $id_order = $_POST['id_order'];
 @ $quantity = $_POST['cantidad'];
 
+if ($quantity > 999) {
+	echo "<script>alert('No se pueden comprar más de 999 unidades de un producto');history.back();</script>";
+	exit;
+}
+
 include('conexBD.php');
+
 
 if($quantity == 0){
     $query= "delete from order_products where id_product = ".$id_product." AND id_pedido = ".$id_order."";
